+++ src/components/Home.js
@@ -1,0 +1,5 @@
+const Home = () => (
+  <div>Home</div>
+);
+
+export default Home;
