+++ src/components/Quote.js
@@ -1,0 +1,5 @@
+const Quote = () => (
+  <div>Quote</div>
+);
+
+export default Quote;
